MAINT: Split logging into its own file. This will be a useful split because more is coming in for the multiprocess logging.
